+++ src/main/java/com/example/kakaoshop/cart/response/ProductDTO.java
@@ -8,12 +8,12 @@ import java.util.List;
 
 @Getter @Setter
 public class ProductDTO {
-    private int id;
+    private Long id;
     private String productName;
     private List<CartItemDTO> cartItems;
 
     @Builder
-    public ProductDTO(int id, String productName, List<CartItemDTO> cartItems) {
+    public ProductDTO(Long id, String productName, List<CartItemDTO> cartItems) {
         this.id = id;
         this.productName = productName;
         this.cartItems = cartItems;
