+++ src/main/java/com/example/kakaoshop/product/response/ProductOptionDTO.java
@@ -7,12 +7,12 @@ import lombok.Setter;
 @Getter @Setter
 public class ProductOptionDTO {
 
-    private int id;
+    private Long id;
     private String optionName;
     private int price;
 
     @Builder
-    public ProductOptionDTO(int id, String optionName, int price) {
+    public ProductOptionDTO(Long id, String optionName, int price) {
         this.id = id;
         this.optionName = optionName;
         this.price = price;
