+++ src/main/java/com/example/kakaoshop/product/response/ProductRespFindAllDTO.java
@@ -7,14 +7,14 @@ import lombok.Setter;
 @Getter @Setter
 public class ProductRespFindAllDTO {
 
-    private int id;
+    private Long id;
     private String productName;
     private String description;
     private String image;
     private int price;
 
     @Builder
-    public ProductRespFindAllDTO(int id, String productName, String description, String image, int price) {
+    public ProductRespFindAllDTO(Long id, String productName, String description, String image, int price) {
         this.id = id;
         this.productName = productName;
         this.description = description;
